+++ next-docs/pages/components/typography/heading.tsx
@@ -17,7 +17,7 @@ const PageHeading = () => {
   return (
     <>
       <ComponentPageDescription title="Heading">
-        <DeprecationWarning href="/core/typography" name="Typography" />
+        <DeprecationWarning href="/typography" name="Typography" />
       </ComponentPageDescription>
       <Preview
         title="Colors"