+++ packages/notifications/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 9.15.0
+
+### Minor Changes
+
+- Version bump
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.15.0
+  - @heathmont/moon-utils@9.15.0
+
 ## 9.14.3
 
 ### Patch Changes