+++ packages/select/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @heathmont/moon-select
 
+## 9.15.0
+
+### Minor Changes
+
+- Version bump
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.15.0
+  - @heathmont/moon-themes@9.15.0
+  - @heathmont/moon-utils@9.15.0
+
 ## 9.14.3
 
 ### Patch Changes