+++ workspaces/tables/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @heathmont/moon-table-tw
 
+## 9.15.0
+
+### Minor Changes
+
+- Version bump
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.15.0
+
 ## 9.14.3
 
 ### Patch Changes