+++ workspaces/templates/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-templates
 
+## 9.15.0
+
+### Minor Changes
+
+- Version bump
+
 ## 9.14.3
 
 ### Patch Changes